refactor: DefinitionGenerator の移動
